refactor:connect users and their tasks, migrate to UID system and add Zod validation

+++ server/src/services/task.sevice.js
@@ -1,26 +1,40 @@
 const prisma = require("../../prisma.js");
 
-const getAllTasks = async () => {
-    return prisma.task.findMany()
+const getAllTasks = async (userId) => {
+    return prisma.task.findMany({
+        where: {userId},
+        orderBy: {id: "desc"}
+    })
 }
 
-const createTask = async (title, description, userId = 1) => {
+const createTask = async (title, description, userId) => {
     return prisma.task.create({
         data: {title, description, userId}
     })
 }
 
-const updateTask = async ( id, data ) => {
+const updateTask = async ( id, data, userId) => {
+    await verifyOwnership(id, userId)
+
     return prisma.task.update({
-        where: { id: id },
-        data: data,
+        where: { id },
+        data
     })
 }
 
-const deleteTask = async ( id ) => {
+const deleteTask = async ( id, userId) => {
+    await verifyOwnership(id, userId)
+
     return prisma.task.delete({
-        where: { id: id }
+        where: { id }
     })
+}
+
+const verifyOwnership = async(taskId, userId) =>{
+    const task = await prisma.task.findUnique({ where: {id: taskId}})
+
+    if(!task) throw new Error("Task not found")
+    if( task.userId !== userId ) throw new Error("No access to task")
 }
 
 module.exports = {
